actualizacion rutas vinculadas con el frontend
